requirements.txt
Django==3.1.5
Pillow==6.1.0
django-import-export==2.1.0
django-rest-swagger==2.2.0
djangorestframework==3.11.0
drf-yasg==1.17.1
django-filter==2.3.0
